chatbot/backend/requirements.txt
fastapi
uvicorn
neo4j
openai
python-dotenv
requests